requirements.txt
pygments-2.11.2
requests-2.28.1